+++ packages/api/src/learning-resources/core/generate-learning-resources.command-handler.ts
@@ -4,7 +4,7 @@ import {GenerateLearningResources} from "../api/generate-learning-resources.comm
 import {LEARNING_RESOURCES_REPOSITORY, LearningResourcesRepository} from "./learning-resources.repository";
 import { Inject } from "@nestjs/common";
 import {LEARNING_RESOURCES_GENERATOR, LearningResourcesGenerator} from "./learning-resources-generator";
-import {TIME_PROVIDER, TimeProvider} from "./time-provider";
+import {TIME_PROVIDER, TimeProviderPort} from "./time-provider";
 import dayjs from 'dayjs'
 
 @CommandHandler(GenerateLearningResources)
@@ -14,7 +14,7 @@ export class GenerateLearningResourcesCommandHandler implements ICommandHandler<
     private readonly eventBus: EventBus,
     @Inject(LEARNING_RESOURCES_REPOSITORY) private readonly repository: LearningResourcesRepository,
     @Inject(LEARNING_RESOURCES_GENERATOR) private readonly learningResourcesGenerator: LearningResourcesGenerator,
-    @Inject(TIME_PROVIDER) private readonly timeProvider: TimeProvider,
+    @Inject(TIME_PROVIDER) private readonly timeProvider: TimeProviderPort,
   ) {
   }
