+++ packages/api/src/learning-resources/core/learning-resources.model.ts
