+++ packages/api/src/learning-resources/core/learning-resources.repository.ts
@@ -1,5 +1,5 @@
 import {UserId} from "../../shared/user-id";
-import {LearningResources} from "./learning.resources";
+import {LearningResources} from "./learning-resources.model";
 
 export const LEARNING_RESOURCES_REPOSITORY = Symbol("LearningResourcesRepository")
 export interface LearningResourcesRepository {
