+++ packages/api/src/learning-resources/infrastructure/system-time-provider.ts
@@ -1,6 +1,6 @@
-import {TimeProvider} from "../core/time-provider";
+import {TimeProviderPort} from "../core/time-provider";
 
-export class SystemTimeProvider implements TimeProvider {
+export class SystemTimeProvider implements TimeProviderPort {
   currentTime(): Date {
     return new Date();
   }
